Removed hide run tab. Split parser into parser and lexer. Old submodule and module work temporarily removed.
