+++ Visiboole/Resources/Sample Files/CLA.vbi
@@ -5,15 +5,15 @@
         g[15..0];
         p[15..0];
         C V G[3..0] P[3..0];
-        
+
         g[] = a[15..0] b[15..0];
         p[] = a[15..0] | b[15..0];
         s[] = a[15..0] ^ b[15..0] ^ c[15..0];
         C = c16;
         V = c16 ^ c15;
-        
-        CLU.b3to0(g[3..0], p[3..0], {NC c[3..0]}, G0, P0);
-        CLU.b7to4(g[7..4], p[7..4], {NC c[7..4]}, G1, P1);
-        CLU.b11to8(g[11..8], p[11..8], {NC c[11..8]}, G2, P2);
-        CLU.b15to12(g[15..12], p[15..12], {NC c[15..12]}, G3, P3);
-        CLU.top(G[3..0], P[3..0], c[16.4.0], NC, NC);
+
+        CLU.b3to0(g[3..0], p[3..0] : {NC c[3..0]}, G0, P0);
+        CLU.b7to4(g[7..4], p[7..4] : {NC c[7..4]}, G1, P1);
+        CLU.b11to8(g[11..8], p[11..8] : {NC c[11..8]}, G2, P2);
+        CLU.b15to12(g[15..12], p[15..12] : {NC c[15..12]}, G3, P3);
+        CLU.top(G[3..0], P[3..0] : c[16.4.0], NC, NC);
