+++ Visiboole/Resources/Sample Files/testing.vbi
@@ -1,0 +1,2 @@
+*d[3..0];
+a[3..0] = ({b3 b2 b1 b0} {c3 c2 c1 c0}) ^ (d[]);
