+++ Visiboole/Resources/Sample Files/testing2.vbi
@@ -1,0 +1,2 @@
+a[4..0];
+a[5..0] = 6'b1;
