+++ Visiboole/Resources/Sample Files/testing3.vbi
@@ -1,0 +1,3 @@
+a[4..0];
+a7;
+a[] = {8'b1};
